+++ src/lib.rs
@@ -8,10 +8,11 @@ mod one;
 mod sqrt;
 mod vec2;
 mod vec2_range_iterator;
+mod zero;
 
 pub use self::{
     cast::*, containing_float::*, float::*, goniometric::*, into_float::*,
-    isqrt::*, one::*, sqrt::*, vec2::*, vec2_range_iterator::*,
+    isqrt::*, one::*, sqrt::*, vec2::*, vec2_range_iterator::*, zero::*,
 };
 
 #[cfg(test)]